backend/requirements.txt
# --- Django core ---
Django==5.2.7
dj-database-url==3.0.1
gunicorn==23.0.0
whitenoise==6.11.0

# --- REST + Auth ---
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==24.1

# --- Database ---
psycopg2-binary==2.9.11
sqlparse==0.5.3

# --- ML / Data ---
numpy==2.3.4
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.16.2
lightgbm==4.6.0
catboost==1.2.8
joblib==1.5.2
holidays==0.84
threadpoolctl==3.6.0

# --- APIs ---
requests==2.32.5
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
openmeteo_requests==1.7.4
openmeteo_sdk==1.22.0
protobuf>=3.20,<6.0.0

# --- Time/Date utils ---
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# --- Environment & Config ---
python-dotenv==1.0.1

# --- Serialization ---
jsonschema==4.25.1
pydantic==2.12.3

# --- Utilities ---
Pillow==11.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0
requests-cache==1.2.1
typing-extensions==4.15.0
